Extract wh, completion, experiential, and scalar questions
